refactor(control): split nfqueue runtime
